# sansa-auth/target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: Project Sansa Auth API
  version: 0.1.0
paths:
  /auth/pre-register:
    post:
      summary: Pre-register by email (send verification code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
              required: [email]
      responses:
        '200':
          description: Sent
  /auth/verify-email:
    post:
      summary: Verify email code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                code: { type: string, minLength: 6, maxLength: 6 }
              required: [email, code]
      responses:
        '200':
          description: PreReg issued
  /auth/register:
    post:
      summary: Finalize registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preRegId: { type: string }
                accountId: { type: string }
                language: { type: string }
              required: [preRegId, accountId]
      responses:
        '200': { description: OK }
  /webauthn/challenge:
    get:
      summary: Begin WebAuthn assertion
      responses:
        '200': { description: OK }
  /webauthn/assertion:
    post:
      summary: Complete WebAuthn assertion (demo accepts accountId only)
      parameters:
        - in: query
          name: accountId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /auth/mfa/totp/verify:
    post:
      summary: Verify TOTP (demo accepts 000000)
      responses:
        '200': { description: OK }
  /auth/mfa/email/request:
    post:
      summary: Request email OTP
      responses:
        '200': { description: OK }
  /auth/mfa/email/verify:
    post:
      summary: Verify email OTP (demo accepts 000000)
      responses:
        '200': { description: OK }
  /auth/logout_all:
    post:
      summary: Logout all devices (token_version++)
      parameters:
        - in: query
          name: accountId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }